MNT: Combine vector transform tests into a single file
Move crs_transform_vectors tests into the test_vector_transforms file.
